updated SQL in Delete Contacts

+++ LAMPAPI/DeleteContacts.php
@@ -12,9 +12,10 @@
 
         // defining vars
         $fName = $inData["firstName"];
-        $lName = $inData["lastName"];
+	$lName = $inData["lastName"];
+	$userID= $inData["userID"];
 
-        $query = "DELETE FROM Contacts WHERE FirstName='$fName', LastName='$lName'";
+	$query = "DELETE FROM Contacts WHERE UserID='$userID' AND FirstName='$fName' AND LastName='$lName'";
         mysqli_query($conn,$query);
 
         $conn->close();
